requirements.txt
selenium==4.21.0
mysql-connector-python==8.4.0
python-dotenv==1.0.1
webdriver_manager==4.0.1
nltk==3.8.1
unidecode==1.3.8